fix: build contract if has not been built
