+++ crates/pop-cli/src/common/mod.rs
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 #[cfg(feature = "contract")]
+pub mod build;
+#[cfg(feature = "contract")]
 pub mod contracts;
 pub mod helpers;
